+++ packages/cubejs-schema-compiler/test/PreAggregationsTest.js
@@ -94,27 +94,13 @@ describe('PreAggregations', function test() {
           measureReferences: [checkinsTotal],
           segmentReferences: [google],
           timeDimensionReference: createdAt,
-          granularity: 'day',
+          granularity: 'week',
         },
         approx: {
           type: 'rollup',
           measureReferences: [countDistinctApprox],
           timeDimensionReference: createdAt,
           granularity: 'day'
-        },
-        ratio: {
-          type: 'rollup',
-          measureReferences: [checkinsTotal, uniqueSourceCount],
-          timeDimensionReference: createdAt,
-          granularity: 'day'
-        },
-        partitioned: {
-          type: 'rollup',
-          measureReferences: [checkinsTotal],
-          dimensionReferences: [source],
-          timeDimensionReference: createdAt,
-          granularity: 'day',
-          partitionGranularity: 'month'
         },
         multiStage: {
           useOriginalSqlPreAggregations: true,
@@ -126,6 +112,20 @@ describe('PreAggregations', function test() {
           refreshKey: {
             sql: \`SELECT CASE WHEN \${FILTER_PARAMS.visitors.createdAt.filter((from, to) => \`\${to}::timestamp > now()\`)} THEN now() END\`
           }
+        },
+        partitioned: {
+          type: 'rollup',
+          measureReferences: [checkinsTotal],
+          dimensionReferences: [source],
+          timeDimensionReference: createdAt,
+          granularity: 'day',
+          partitionGranularity: 'month'
+        },
+        ratio: {
+          type: 'rollup',
+          measureReferences: [checkinsTotal, uniqueSourceCount],
+          timeDimensionReference: createdAt,
+          granularity: 'day'
         }
       }
     })
@@ -518,7 +518,7 @@ describe('PreAggregations', function test() {
         preAggregationsSchema: '',
         timeDimensions: [{
           dimension: 'visitors.createdAt',
-          granularity: 'date',
+          granularity: 'week',
           dateRange: ['2016-12-30', '2017-01-05']
         }],
         order: [{
@@ -542,7 +542,7 @@ describe('PreAggregations', function test() {
         res.should.be.deepEqual(
           [
             {
-              "visitors__created_at_date": "2017-01-05T00:00:00.000Z",
+              "visitors__created_at_week": "2017-01-02T00:00:00.000Z",
               "visitors__checkins_total": "1"
             }
           ]
@@ -584,6 +584,160 @@ describe('PreAggregations', function test() {
             { "visitors__source": "some", "visitors__checkins_total": "5" },
             { "visitors__source": "google", "visitors__checkins_total": "1" },
             { "visitors__source": null, "visitors__checkins_total": "0" }
+          ]
+        );
+      });
+    });
+  });
+});
+
+
+describe('PreAggregations in time hierarchy', function test() {
+  this.timeout(20000);
+
+  after(async () => {
+    await dbRunner.tearDown();
+  });
+
+  const { compiler, joinGraph, cubeEvaluator } = prepareCompiler(`
+    cube(\`visitors\`, {
+      sql: \`
+      select * from visitors
+      \`,
+
+      measures: {
+        count: {
+          type: 'count'
+        }
+      },
+
+      dimensions: {
+        createdAt: {
+          type: 'time',
+          sql: 'created_at'
+        },
+      },
+      
+      preAggregations: {
+        month: {
+          type: 'rollup',
+          measureReferences: [count],
+          timeDimensionReference: createdAt,
+          granularity: 'month',
+        },
+        day: {
+          type: 'rollup',
+          measureReferences: [count],
+          timeDimensionReference: createdAt,
+          granularity: 'day',
+        },
+      }
+    })
+    `);
+
+  function replaceTableName(query, preAggregation, suffix) {
+    const [toReplace, params] = query;
+    console.log(toReplace);
+    preAggregation = Array.isArray(preAggregation) ? preAggregation : [preAggregation];
+    return [
+      preAggregation.reduce((replacedQuery, desc) =>
+        replacedQuery.replace(new RegExp(desc.tableName, 'g'), desc.tableName + '_' + suffix), toReplace
+      ),
+      params
+    ];
+  }
+
+  function tempTablePreAggregations(preAggregationsDescriptions) {
+    return R.unnest(preAggregationsDescriptions.map(desc =>
+      desc.invalidateKeyQueries.concat([
+        [desc.loadSql[0].replace('CREATE TABLE', 'CREATE TEMP TABLE'), desc.loadSql[1]]
+      ])
+    ));
+  }
+
+  it('query on year match to pre-agg on month', () => {
+    return compiler.compile().then(() => {
+      const query = new PostgresQuery({ joinGraph, cubeEvaluator, compiler }, {
+        measures: [
+          'visitors.count'
+        ],
+        dimensions: [],
+        timezone: 'America/Los_Angeles',
+        timeDimensions: [{
+          dimension: 'visitors.createdAt',
+          granularity: 'year',
+          dateRange: ['2016-12-30', '2018-12-30']
+        }],
+        preAggregationsSchema: '',
+        order: [],
+      });
+
+      const queryAndParams = query.buildSqlAndParams();
+
+      query.preAggregations.preAggregationForQuery.preAggregation.granularity.should.be.equal('month');
+
+      console.log(queryAndParams);
+      const preAggregationsDescription = query.preAggregations.preAggregationsDescription();
+      console.log(preAggregationsDescription);
+
+      const queries = tempTablePreAggregations(preAggregationsDescription);
+
+      console.log(JSON.stringify(queries.concat(queryAndParams)));
+
+      return dbRunner.testQueries(
+        queries.concat([queryAndParams]).map(q => replaceTableName(q, preAggregationsDescription, 1))
+      ).then(res => {
+        console.log(JSON.stringify(res));
+        res.should.be.deepEqual(
+          [
+            {
+              "visitors__count": "5",
+              "visitors__created_at_year": "2017-01-01T00:00:00.000Z"
+            },
+          ]
+        );
+      });
+    });
+  });
+  it('query on week match to pre-agg on day', () => {
+    return compiler.compile().then(() => {
+      const query = new PostgresQuery({ joinGraph, cubeEvaluator, compiler }, {
+        measures: [
+          'visitors.count'
+        ],
+        dimensions: [],
+        timezone: 'America/Los_Angeles',
+        timeDimensions: [{
+          dimension: 'visitors.createdAt',
+          granularity: 'week',
+          dateRange: ['2017-01-02', '2019-02-08']
+        }],
+        preAggregationsSchema: '',
+        order: [],
+      });
+
+      const queryAndParams = query.buildSqlAndParams();
+
+      query.preAggregations.preAggregationForQuery.preAggregation.granularity.should.be.equal('day');
+
+      console.log(queryAndParams);
+      const preAggregationsDescription = query.preAggregations.preAggregationsDescription();
+      console.log(preAggregationsDescription);
+
+      const queries = tempTablePreAggregations(preAggregationsDescription);
+
+      console.log(JSON.stringify(queries.concat(queryAndParams)));
+
+      return dbRunner.testQueries(
+        queries.concat([queryAndParams]).map(q => replaceTableName(q, preAggregationsDescription, 1))
+      ).then(res => {
+        console.log(JSON.stringify(res));
+        res.should.be.deepEqual(
+          [
+            {
+              "visitors__count": "5",
+              "visitors__created_at_week": "2017-01-02T00:00:00.000Z"
+            },
           ]
         );
       });