+++ packages/cubejs-schema-compiler/test/SQLGenerationTest.js
@@ -28,8 +28,6 @@ describe('SQL Generation', function test() {
       \${USER_CONTEXT.sourceArray.filter(sourceArray => \`source in (\${sourceArray.join(',')})\`)}
       \`,
       
-      sqlAlias: 'visitors_table',
-      
       refreshKey: {
         sql: 'SELECT 1',
       },
@@ -101,7 +99,8 @@ describe('SQL Generation', function test() {
         averageCheckins: {
           type: 'avg',
           sql: \`\${doubledCheckings}\`
-        }
+        },
+        ...(['foo', 'bar'].map(m => ({ [m]: { type: 'count' } })).reduce((a, b) => ({ ...a, ...b })))
       },
 
       dimensions: {
@@ -288,6 +287,22 @@ describe('SQL Generation', function test() {
         }
       }
     })
+    
+    cube('CubeWithVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryLongName', {
+      sql: \`
+      select * from cards
+      \`,
+      
+      sqlAlias: 'cube_with_long_name',
+      
+      dataSource: 'oracle',
+
+      measures: {
+        count: {
+          type: 'count'
+        }
+      }
+    });
     `);
 
   it('simple join', () => {
@@ -1400,4 +1415,45 @@ describe('SQL Generation', function test() {
   }, {
     "visitors__created_at_date": "2017-01-06T00:00:00.000Z"
   }]));
+
+  it(
+    'sqlAlias',
+    () => runQueryTest({
+      measures: ['CubeWithVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryLongName.count'],
+      dimensions: [],
+      timeDimensions: [],
+      timezone: 'America/Los_Angeles',
+      filters: [],
+      order: []
+    }, [
+      { "cube_with_long_name__count": '3' }
+    ])
+  );
+
+  it('data source', () => compiler.compile().then(() => {
+    const query = new PostgresQuery({ joinGraph, cubeEvaluator, compiler }, {
+      measures: ['CubeWithVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryVeryLongName.count'],
+      dimensions: [],
+      timeDimensions: [],
+      timezone: 'America/Los_Angeles',
+      filters: [],
+      order: []
+    });
+
+    query.dataSource.should.be.deepEqual('oracle');
+  }));
+
+  it(
+    'objectRestSpread generator',
+    () => runQueryTest({
+      measures: ['visitors.foo'],
+      dimensions: [],
+      timeDimensions: [],
+      timezone: 'America/Los_Angeles',
+      filters: [],
+      order: []
+    }, [
+      { "visitors__foo": '6' }
+    ])
+  );
 });